requirements.txt
allure-pytest==2.8.16
flask==1.1.2
gunicorn==19.9.0
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.1
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
cloudpickle==1.2.2
ConfigSpace==0.4.12
cycler==0.10.0
Cython==0.29.15
decorator==4.4.2
defusedxml==0.6.0
distributed==2.12.0
docutils==0.16
eli5==0.10.1
entrypoints==0.3
fire==0.2.1
fsspec==0.6.3
future==0.18.2
google-api-core==1.14.3
google-auth==1.7.1
google-cloud-core==1.0.3
google-cloud-translate==2.0.0
googleapis-common-protos==1.6.0
grpcio==1.25.0
HeapDict==1.0.1
hyperopt==0.2.2
idna==2.8
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.2
joblib==0.14.0
json5==0.8.5
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
jupyterlab==1.2.0
jupyterlab-server==1.0.6
kiwisolver==1.1.0
liac-arff==2.4.0
llvmlite==0.31.0
locket==0.2.0
lockfile==0.12.2
lxml==4.4.1
MarkupSafe==1.1.1
matplotlib==3.2.0
mistune==0.8.4
more-itertools==8.2.0
msgpack==1.0.0
networkx==2.2
nltk==3.4.5
nose==1.3.7
numpy==1.18.2
packaging==20.3
pandas==1.0.2
pandocfilters==1.4.2
parso==0.6.2
partd==1.1.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.10.0
ptvsd==4.3.2
ptyprocess==0.6.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
Pygments==2.6.1
pynisher==0.5.0
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.0
requests==2.22.0
retrying==1.3.3
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
sortedcontainers==2.1.0
swifter==0.301
tabulate==0.8.6
tblib==1.6.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toolz==0.10.0
tornado==6.0.4
tqdm==4.38.0
traitlets==4.3.3
typing-extensions==3.7.4.1
urllib3==1.25.7
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
zict==2.0.0
zipp==3.1.0